+++ changelog.md
@@ -11,6 +11,7 @@
 - Removed various short TNM false positives (e.g., "PT" or "a   T")
 - The Span value extension is not more forcibly overwritten, and user assigned values are returned by `Span._.value` in priority, before the aggregated `span._.get(span.label_)` getter result (#220)
 - Enable mmap during multiprocessing model transfers
+- `RegexMatcher` now supports all alignment modes (`strict`, `expand`, `contract`) and better handles partial doc matching (#201).
 
 ## v0.10.0